path planner
the function that adds a new points is running really quick, cause it is based on evaluating where along the current route the new points should be inserted based on ONLY distances to the other waypoints and the current state. The dyn_planner is taking into account the dynamics of the plane using the controller and stats.